+++ inconnu/models/__init__.py
@@ -1,6 +1,6 @@
 """Database model package."""
 
 from inconnu.models.manager import CharacterManager
-from inconnu.models.post import RPPost
 from inconnu.models.rpheader import HeaderSubdoc
+from inconnu.models.rppost import RPPost
 from inconnu.models.vchar import VChar
